# Avoid LazilyDefinedAttributes being included multiple times in class, and remove methods after first use