Solution to issues in #35968 and #35966. Also fixes #34659.

### Summary

Currently, the acceptance validator includes an anonymous module in order to override `method_missing` and `respond_to_missing?` so that the validator can catch calls to attribute methods and define attribute accessors lazily (to avoid a schema load).

The problem with this is that each call to `validates` on a model will include another module into the class. I've already fixed a bunch of tests to avoid polluting test classes as a result of this, but the underlying problem is still there.

This PR gets around the issue using a normal module and grabbing the attributes from the validators on the model, which are accessible within method_missing and respond_to_missing?

@kamipo If this seems like a reasonable solution, I'd like to revert at least #35968 before merging.