@shioyama 2019-04-14
Also: `def` is faster than `define_method`, so this should have a small performance improvement for anything going through `method_missing` or `respond_to?`.