@kamipo 2019-04-14
Reverting #35968 is less worth to me, but at least the comment should be removed, so either is fine.

https://github.com/rails/rails/blob/f8944ed22dc9453dea5814676fa49ce119a58726/activemodel/test/cases/validations/acceptance_validation_test.rb#L96-L98

And also, this fixes #34659 fortunately, can you add a test case for #34659?