@shioyama 2019-04-14
> And also, this fixes #34659 fortunately, can you add a test case for #34659?

Oh wow, what a pleasant surprise! I'll add a test case later tonight. Thanks for quick response!