@kamipo 2019-04-14
> Also: `def` is faster than `define_method`, so this should have a small performance improvement for anything going through `method_missing` or `respond_to?`, not to mention that if you've added acceptance validators multiple times, it will cut the number of overrides from N to just one.

I don't know this have a performance improvement, in exchange for `define_method` to `def`, this makes each `matches?` generates extra an array and strings.