@shioyama 2019-04-14
cdb035c is an attempt to minimize the number of string conversions (partially). It seems that `attribute_method?` takes a symbol or string.