@shioyama 2019-04-14
> I don't know this have a performance improvement, in exchange for define_method to def, this makes each matches? generates extra an array and strings.

Good point. `matches?` can actually stop once it finds a symbol that converts to a matching string, so I improved that a bit.