@kamipo 2019-04-14
`record.respond_to?(:unknown_method)` is perhaps slower though.