@shioyama 2019-04-14
Added a quick test in be91bd9.