@shioyama 2019-04-14
Wow noticed there are yet *more* places where things are not being cleaned up after acceptance validators are added to classes:

https://github.com/shioyama/rails/blob/9951f91a12b71f97cab6ea090109a1b37a1ab5d4/activemodel/test/cases/validations/i18n_validation_test.rb#L187-L193

This stuff makes it hard to test changes since totally unrelated specs start breaking... we need to ensure all tests with acceptance validators use subclasses to avoid polluting. Even in case where there is only one module.