@shioyama 2019-04-14
It just occurred to me that activemodel validations are supposed to work on any class that includes `ActiveModel::Validations`, but the acceptance validator implicitly requires `ActiveModel::AttributeMethods` when it calls `define_attribute_methods`...