@kamipo 2019-04-14
Calling `define_attribute_methods` is added at ce48b5a, to prevent infinite method_missing loop.
Add `klass.respond_to?(:define_attribute_methods)` is fine to me.

`attribute_method?` is a helper method for validations.

https://github.com/rails/rails/blob/9c87474d82165277a6a5ed3e04b6b88402968134/activemodel/lib/active_model/validations.rb#L270-L272