@shioyama 2019-04-14
Funny you mention ce48b5a, reverting that change (well, removing that line, I didn't check the git blame) was what got me to this pull request in the first place... I also saw the infinite loop but all else equal, it would be better not to trigger that in a validator override... but now I'm getting stuck.

The basic problem here is that this validator needs to work with both ActiveModel and ActiveRecord. Maybe things would be easier if activerecord just had something custom... otherwise, I think my current solution here is the cleanest one we're going to get.