@shioyama 2019-04-14
What do you think about just defining an activerecord acceptance validation that subclasses this one?