@shioyama 2019-04-14
Ok in #35971 I fixed the additional pollution problems mentioned above. I realized that even with the change here, it's still an issue so everything in tests should use subclasses not original test classes.