@shioyama 2019-04-15
@kamipo how about d44fb37 ? `dup` the module, attach it to the model namespace, then once accessors have been defined, undefine `method_missing` and `respond_to_missing?` from it. That means there is zero performance hit after the first model instance defines any acceptance method.

I think also that this should allow us to remove `define_attribute_methods`, since once the module is removed, there should be no infinite cycle. But since there's no test for that I haven't had a chance to check.