@shioyama 2019-04-15
@kamipo This seems to be working now.

Basically:
- we `dup` the original module and name it under the model class, then include it
- anytime we hit `method_missing`, we immediately define acceptance accessors and "implode" the module by removing both `method_missing` and `respond_to_missing?` on it, since we don't need them anymore.

So basically, with this setup:
- only the first instance to hit `method_missing` takes a performance hit on that method, after that the module has no effect (since we've removed the method)
- `respond_to_missing?` will still have a performance hit until we hit `method_missing`, but that should be fixable in the same way (test is failing now on that so haven't committed yet, will investigate)

I'm not an expert on the implications of using `remove_method` but I don't see why this should be an issue. Also, I'm pretty sure this will fix the infinite loop so we shouldn't need that `define_attribute_methods`, but we'll need a test for that first.

What do you think?