@shioyama 2019-04-15
Ok I've rewritten as a module builder again (as it was originally), but in such a way that there is never more than one. This passes specs for me (some unrelated failures on buildkite, not sure what's up), and should be faster and cleaner than the current solution. It also removes the `define_attribute_methods`, and passes the spec for that which I noticed was added in d99984d48c6.

So I'm happy with this. I realized removing methods on a module is maybe not standard but if it works, why not right? You can't impact performance if you're not there!