@shioyama 2019-04-16
Some performance results:

```ruby
# frozen_string_literal: true

require "active_record"
require "benchmark/ips"

conn = { adapter: "sqlite3", database: ":memory:" }

ActiveRecord::Base.establish_connection(conn)

class Person < ActiveRecord::Base
  connection.create_table :people, force: :true do |t|
    t.string :name
  end

  validates :terms_of_service, acceptance: true
  validates :eula, acceptance: true
end

class Topic
  include ActiveModel::Validations
  include ActiveModel::Validations::Callbacks
  include ActiveModel::AttributeMethods
  include ActiveSupport::NumberHelper

  validates :terms_of_service, acceptance: true
  validates :eula, acceptance: true

  attr_reader :attributes
  
  def initialize(*)
    @attributes = { 'foo' => nil }
  end

  def attribute(*)
  end
end

Person.create

Benchmark.ips(10) do |x|
  person = Person.new
  foo_person = Person.select('name as foo').first
  topic = Topic.new

  x.report('AR respond_to?') do
    person.respond_to?('miss')
  end

  x.report('AR method_missing') do
    foo_person.foo
  end

  x.report('AM respond_to?') do
    topic.respond_to?('miss')
  end

  x.report('AM method_missing') do
    topic.foo
  end
end
```

**BEFORE**

```
      AR respond_to?    403.265k (± 0.5%) i/s -      4.043M in  10.025001s
   AR method_missing    304.773k (± 0.9%) i/s -      3.055M in  10.024724s
      AM respond_to?    452.244k (± 2.0%) i/s -      4.540M in  10.042971s
   AM method_missing    381.810k (± 1.0%) i/s -      3.852M in  10.090269s
```

**AFTER**

```
      AR respond_to?    865.161k (± 0.5%) i/s -      8.716M in  10.074354s
   AR method_missing    576.058k (± 1.0%) i/s -      5.792M in  10.055779s
      AM respond_to?      1.153M (± 0.9%) i/s -     11.544M in  10.013341s
   AM method_missing    934.458k (± 0.8%) i/s -      9.368M in  10.025936s
```

So with two acceptance attributes, this **more than doubles the performance** of `respond_to?` and `method_missing`. (In the example, I use `select` on the AR model to ensure that there is something hit with `method_missing` (to avoid an exception), and ensure ther is a `foo` key on the AM model for the same reason).

The reason is simple: the changes here remove the overrides after first access on any instance, which happens almost instantaneously, and after that they have zero impact, so the "after" results are basically what you get with no acceptance validation at all.